+++ custom-recipes/regression-splines/recipe.py
@@ -6,7 +6,7 @@ from dataiku import pandasutils as pdu
 
 
 from regression_splines.dku_reg_splines import RegressionSplines
-from regression_splines.commons import *
+from commons import *
 # define inputs
 (input_dataset, output_dataset) = get_input_output()
 recipe_config = get_recipe_config()
@@ -21,7 +21,7 @@ df = input_dataset.get_dataframe()
 
 
 RegSpines = RegressionSplines(column_name, degree_freedom, knots, new_col_prefix)
-output_dataset_df = RegSpines.run_spline_creation(df,keep_original=True)
+output_dataset_df = RegSpines.run_spline_creation(df, keep_original=True)
 
 # Write recipe outputs
 output_dataset.write_with_schema(output_dataset_df)
